feat(AttendanceRecord): 기준 시간과 출석 시간을 기준으로 출석 상태를 판별하는 기능 구현
 - 월요일 기준 시각(13:00)과의 시간 차이를 이용해 출석, 지각, 결석 상태 판별

+++ src/main/java/attendance/domain/AttendanceStatus.java
@@ -8,14 +8,14 @@ public enum AttendanceStatus {
     ABSENT("결석", Integer.MAX_VALUE);
 
     private final String title;
-    private final int limitTime;
+    private final long limitTime;
 
-    AttendanceStatus(String title, int limitTime) {
+    AttendanceStatus(String title, long limitTime) {
         this.title = title;
         this.limitTime = limitTime;
     }
 
-    public static AttendanceStatus from(int lateTime) {
+    public static AttendanceStatus from(long lateTime) {
         return Arrays.stream(AttendanceStatus.values())
                 .filter(status -> lateTime <= status.limitTime)
                 .findFirst()
